Sync hyprland clients on group operations and activewindow.
On group operations positions of clients change, both because of
grouping and the addition of the group tabs.

As for moving a window or resizing it, these are signaled only via
activewindowv2 signal, in which case we need once again to recalculate
positions and maybe even the client monitor.

TODO in another PR: add group information to the hyprland client so that
apps can know if a client is in a group or not.

Closes https://github.com/Aylur/astal/issues/365